TelegramBot_Deployment/Flask Heroku/requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
future==0.18.2
gast==0.3.3
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mtcnn==0.1.0
numpy==1.18.1
opencv-python==4.1.2.30
opencv-python-headless==4.1.2.30
protobuf==3.11.2
pycparser==2.19
python-telegram-bot==12.3.0
PyYAML==5.3
requests==2.22.0
scipy==1.4.1
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==6.0.3
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.11.2
